Refactor binary loading
