+++ kernel/source/Process.c
@@ -13,6 +13,11 @@
 #include "../include/Console.h"
 #include "../include/File.h"
 #include "../include/Kernel.h"
+#include "../include/binary/Binary-EXOS.h"
+#include "../include/binary/Binary-ELF.h"
+
+#define BINARY_TYPE_EXOS 1
+#define BINARY_TYPE_ELF  2
 
 /***************************************************************************/
 
@@ -54,254 +59,6 @@ void InitializeKernelHeap() {
     ((LPHEAPCONTROLBLOCK)KernelProcess.HeapBase)->ID = ID_HEAP;
 }
 
-/***************************************************************************/
-
-BOOL GetExecutableInfo_EXOS(LPFILE File, LPEXECUTABLEINFO Info) {
-    FILEOPERATION FileOperation;
-    EXOSCHUNK Chunk;
-    EXOSHEADER Header;
-    EXOSCHUNK_INIT Init;
-    U32 BytesRead;
-    U32 Index;
-    U32 Dummy;
-
-#ifdef __DEBUG__
-    KernelPrint("Entering GetExecutableInfo_EXOS\n");
-#endif
-
-    if (File == NULL) return FALSE;
-    if (Info == NULL) return FALSE;
-
-    FileOperation.Size = sizeof(FILEOPERATION);
-    FileOperation.File = (HANDLE)File;
-
-    //-------------------------------------
-    // Read the header
-
-    FileOperation.NumBytes = sizeof(EXOSHEADER);
-    FileOperation.Buffer = (LPVOID)&Header;
-    BytesRead = ReadFile(&FileOperation);
-
-    if (Header.Signature != EXOS_SIGNATURE) {
-#ifdef __DEBUG__
-        KernelPrint("GetExecutableInfo_EXOS() : Bad signature (%08X)\n",
-                    Header.Signature);
-#endif
-
-        goto Out_Error;
-    }
-
-    while (1) {
-        FileOperation.NumBytes = sizeof(EXOSCHUNK);
-        FileOperation.Buffer = (LPVOID)&Chunk;
-        BytesRead = ReadFile(&FileOperation);
-
-        if (BytesRead != sizeof(EXOSCHUNK)) break;
-
-        if (Chunk.ID == EXOS_CHUNK_INIT) {
-            FileOperation.NumBytes = sizeof(EXOSCHUNK_INIT);
-            FileOperation.Buffer = (LPVOID)&Init;
-            BytesRead = ReadFile(&FileOperation);
-
-            if (BytesRead != sizeof(EXOSCHUNK_INIT)) goto Out_Error;
-
-            Info->EntryPoint = Init.EntryPoint;
-            Info->CodeBase = Init.CodeBase;
-            Info->DataBase = Init.DataBase;
-            Info->CodeSize = Init.CodeSize;
-            Info->DataSize = Init.DataSize;
-            Info->StackMinimum = Init.StackMinimum;
-            Info->StackRequested = Init.StackRequested;
-            Info->HeapMinimum = Init.HeapMinimum;
-            Info->HeapRequested = Init.HeapRequested;
-
-            goto Out_Success;
-        } else {
-            for (Index = 0; Index < Chunk.Size; Index++) {
-                FileOperation.NumBytes = 1;
-                FileOperation.Buffer = (LPVOID)&Dummy;
-                BytesRead = ReadFile(&FileOperation);
-            }
-        }
-    }
-
-Out_Success:
-
-#ifdef __DEBUG__
-    KernelPrint("Exiting GetExecutableInfo_EXOS (Success)\n");
-#endif
-
-    return TRUE;
-
-Out_Error:
-
-#ifdef __DEBUG__
-    KernelPrint("Exiting GetExecutableInfo_EXOS (Error)\n");
-#endif
-
-    return FALSE;
-}
-
-/***************************************************************************/
-
-BOOL LoadExecutable_EXOS(LPFILE File, LPEXECUTABLEINFO Info, LINEAR CodeBase,
-                         LINEAR DataBase) {
-    FILEOPERATION FileOperation;
-    EXOSCHUNK Chunk;
-    EXOSHEADER Header;
-    EXOSCHUNK_FIXUP Fixup;
-    LINEAR ItemAddress;
-    U32 BytesRead;
-    U32 Index;
-    U32 CodeRead;
-    U32 DataRead;
-    U32 CodeOffset;
-    U32 DataOffset;
-    U32 NumFixups;
-    U32 Dummy;
-    U32 c;
-
-#ifdef __DEBUG__
-    KernelPrint("Entering LoadExecutable_EXOS\n");
-#endif
-
-    if (File == NULL) return FALSE;
-
-    FileOperation.Size = sizeof(FILEOPERATION);
-    FileOperation.File = (HANDLE)File;
-
-    CodeRead = 0;
-    DataRead = 0;
-
-    CodeOffset = CodeBase - Info->CodeBase;
-    DataOffset = DataBase - Info->DataBase;
-
-#ifdef __DEBUG__
-    KernelPrint("LoadExecutable_EXOS() : CodeBase = %08X\n", CodeBase);
-    KernelPrint("LoadExecutable_EXOS() : DataBase = %08X\n", DataBase);
-#endif
-
-    //-------------------------------------
-    // Read the header
-
-    FileOperation.NumBytes = sizeof(EXOSHEADER);
-    FileOperation.Buffer = (LPVOID)&Header;
-    BytesRead = ReadFile(&FileOperation);
-
-    if (Header.Signature != EXOS_SIGNATURE) {
-        goto Out_Error;
-    }
-
-    while (1) {
-        FileOperation.NumBytes = sizeof(EXOSCHUNK);
-        FileOperation.Buffer = (LPVOID)&Chunk;
-        BytesRead = ReadFile(&FileOperation);
-
-        if (BytesRead != sizeof(EXOSCHUNK)) break;
-
-        if (Chunk.ID == EXOS_CHUNK_CODE) {
-            if (CodeRead == 1) {
-                //-------------------------------------
-                // Only one code chunk allowed
-
-                goto Out_Error;
-            }
-
-#ifdef __DEBUG__
-            KernelPrint("LoadExecutable_EXOS() : Reading code\n");
-#endif
-
-            FileOperation.NumBytes = Chunk.Size;
-            FileOperation.Buffer = (LPVOID)CodeBase;
-            BytesRead = ReadFile(&FileOperation);
-
-            if (BytesRead != Chunk.Size) goto Out_Error;
-
-            CodeRead = 1;
-        } else if (Chunk.ID == EXOS_CHUNK_DATA) {
-            if (DataRead == 1) {
-                //-------------------------------------
-                // Only one data chunk allowed
-
-                goto Out_Error;
-            }
-
-#ifdef __DEBUG__
-            KernelPrint("LoadExecutable_EXOS() : Reading data\n");
-#endif
-
-            FileOperation.NumBytes = Chunk.Size;
-            FileOperation.Buffer = (LPVOID)DataBase;
-            BytesRead = ReadFile(&FileOperation);
-
-            if (BytesRead != Chunk.Size) goto Out_Error;
-
-            DataRead = 1;
-        } else if (Chunk.ID == EXOS_CHUNK_FIXUP) {
-            FileOperation.NumBytes = sizeof(U32);
-            FileOperation.Buffer = (LPVOID)&NumFixups;
-            BytesRead = ReadFile(&FileOperation);
-
-            if (BytesRead != sizeof(U32)) goto Out_Error;
-
-#ifdef __DEBUG__
-            KernelPrint("LoadExecutable_EXOS() : Reading relocations\n");
-#endif
-
-            for (c = 0; c < NumFixups; c++) {
-                FileOperation.NumBytes = sizeof(EXOSCHUNK_FIXUP);
-                FileOperation.Buffer = (LPVOID)&Fixup;
-                BytesRead = ReadFile(&FileOperation);
-
-                if (BytesRead != sizeof(EXOSCHUNK_FIXUP)) goto Out_Error;
-
-                if (Fixup.Section & EXOS_FIXUP_SOURCE_CODE) {
-                    ItemAddress = CodeBase + (Fixup.Address - Info->CodeBase);
-                } else if (Fixup.Section & EXOS_FIXUP_SOURCE_DATA) {
-                    ItemAddress = DataBase + (Fixup.Address - Info->DataBase);
-                } else {
-                    ItemAddress = NULL;
-                }
-
-                if (ItemAddress != NULL) {
-                    if (Fixup.Section & EXOS_FIXUP_DEST_CODE) {
-                        *((U32*)ItemAddress) += CodeOffset;
-                    } else if (Fixup.Section & EXOS_FIXUP_DEST_DATA) {
-                        *((U32*)ItemAddress) += DataOffset;
-                    }
-                }
-            }
-
-            goto Out_Success;
-        } else {
-            for (Index = 0; Index < Chunk.Size; Index++) {
-                FileOperation.NumBytes = 1;
-                FileOperation.Buffer = (LPVOID)&Dummy;
-                BytesRead = ReadFile(&FileOperation);
-            }
-        }
-    }
-
-    if (CodeRead == 0) goto Out_Error;
-
-Out_Success:
-
-#ifdef __DEBUG__
-    KernelPrint("Exiting LoadExecutable_EXOS\n");
-#endif
-
-    return TRUE;
-
-Out_Error:
-
-#ifdef __DEBUG__
-    KernelPrint("Exiting LoadExecutable_EXOS\n");
-#endif
-
-    return FALSE;
-}
-
-/***************************************************************************/
 
 LPPROCESS NewProcess() {
     LPPROCESS This = NULL;
@@ -546,6 +303,7 @@ BOOL CreateProcess(LPPROCESSINFO Info) {
     EXECUTABLEINFO ExecutableInfo;
     TASKINFO TaskInfo;
     FILEOPENINFO FileOpenInfo;
+    FILEOPERATION FileOperation;
     LPPROCESS Process = NULL;
     LPTASK Task = NULL;
     LPFILE File = NULL;
@@ -559,6 +317,8 @@ BOOL CreateProcess(LPPROCESSINFO Info) {
     U32 HeapSize = 0;
     U32 StackSize = 0;
     U32 TotalSize = 0;
+    U32 BinaryType = 0;
+    U32 Signature = 0;
     BOOL Result = FALSE;
 
 #ifdef __DEBUG__
@@ -594,10 +354,41 @@ BOOL CreateProcess(LPPROCESSINFO Info) {
 #endif
 
     //-------------------------------------
+    // Detect binary type
+
+    FileOperation.Size = sizeof(FILEOPERATION);
+    FileOperation.File = (HANDLE)File;
+    FileOperation.NumBytes = sizeof(U32);
+    FileOperation.Buffer = (LPVOID)&Signature;
+    ReadFile(&FileOperation);
+
+    if (Signature == EXOS_SIGNATURE) {
+        BinaryType = BINARY_TYPE_EXOS;
+    } else if (Signature == ELF_SIGNATURE) {
+        BinaryType = BINARY_TYPE_ELF;
+    } else {
+        CloseFile(File);
+        return FALSE;
+    }
+
+    CloseFile(File);
+
+    //-------------------------------------
     // Get executable information
 
-    if (GetExecutableInfo_EXOS(File, &ExecutableInfo) == FALSE) {
-        return FALSE;
+    File = OpenFile(&FileOpenInfo);
+    if (File == NULL) return FALSE;
+
+    if (BinaryType == BINARY_TYPE_EXOS) {
+        if (GetExecutableInfo_EXOS(File, &ExecutableInfo) == FALSE) {
+            CloseFile(File);
+            return FALSE;
+        }
+    } else {
+        if (GetExecutableInfo_ELF(File, &ExecutableInfo) == FALSE) {
+            CloseFile(File);
+            return FALSE;
+        }
     }
 
     CloseFile(File);
@@ -697,8 +488,16 @@ BOOL CreateProcess(LPPROCESSINFO Info) {
     KernelPrint("CreateProcess() : Loading executable...\n");
 #endif
 
-    if (LoadExecutable_EXOS(File, &ExecutableInfo, (LINEAR)CodeBase,
-                            (LINEAR)DataBase) == FALSE) {
+    BOOL LoadResult;
+    if (BinaryType == BINARY_TYPE_EXOS) {
+        LoadResult = LoadExecutable_EXOS(File, &ExecutableInfo, (LINEAR)CodeBase,
+                                         (LINEAR)DataBase);
+    } else {
+        LoadResult = LoadExecutable_ELF(File, &ExecutableInfo, (LINEAR)CodeBase,
+                                        (LINEAR)DataBase);
+    }
+
+    if (LoadResult == FALSE) {
 #ifdef __DEBUG__
         KernelPrint("CreateProcess() : Load failed !\n");
 #endif
